@@ -1,13 +1,14 @@
 using System.Linq;
 using System.Net.Mime;
 using BlazorDemo.Data;
-using Microsoft.AspNetCore.Blazor.Server;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Components.Server;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.ResponseCompression;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace BlazorDemo.Server
 {
@@ -15,21 +16,21 @@ namespace BlazorDemo.Server
     {
         private IConfiguration Configuration { get; }
 
-        public Startup(IHostingEnvironment env, IConfiguration config)
+        public Startup(IWebHostEnvironment env, IConfiguration config)
         {
             Configuration = config;
         }
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc();
+            services.AddMvc().AddNewtonsoftJson();
 
             services.AddResponseCompression(options =>
             {
                 options.MimeTypes = ResponseCompressionDefaults.MimeTypes.Concat(new[]
                 {
                     MediaTypeNames.Application.Octet,
-                    WasmMediaTypeNames.Application.Wasm,
+                    WasmMediaTypeNames.Application.Wasm
                 });
             });
 
@@ -39,13 +40,14 @@ namespace BlazorDemo.Server
             });
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             app.UseResponseCompression();
 
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseBlazorDebugging();
             }
 
             using (var serviceScope = app.ApplicationServices.GetService<IServiceScopeFactory>().CreateScope())
@@ -54,10 +56,12 @@ namespace BlazorDemo.Server
                 context.Database.EnsureCreated();
             }
 
-            app.UseMvc(routes =>
+            app.UseRouting();
+
+            app.UseEndpoints(routes =>
             {
-                routes.MapRoute(name: "default", template: "{controller}/{action}/{id?}");
-                routes.MapRoute(name: "paged", template: "{controller}/{action}/page/{page}");
+                routes.MapDefaultControllerRoute();
+                routes.MapControllerRoute("Paged", "{controller=Home}/{action=Index}/page/{page=1}");
             });
 
             app.UseBlazor<Client.Program>();
