@@ -1,6 +1,5 @@
 ﻿@page "/"
-@page "/page/{Page}"
-@page "/page/{Page}/{SearchTerm}"
+@page "/page/{page}"
 @inherits IndexModel
 
 <div class="row">
